LicenseBoard/requirements.txt
beautifulsoup4==4.8.1
bootstrap4==0.1.0
dj-database-url==0.5.0
Django==2.2.6
django-bootstrap4==1.0.1
django-debug-toolbar==2.0
django-decouple==2.1
django-widget-tweaks==1.4.5
Markdown==3.1.1
packaging==19.2
pip-review==1.0
pyparsing==2.4.2
python-decouple==3.1
pytz==2019.3
six==1.12.0
soupsieve==1.9.4
sqlparse==0.3.0
